TMS Trip Controller Code Review fix
